moved getWfsFeatureProperties to WfsUtils
